backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
httpx==0.25.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.1
python-dotenv==1.0.0 
pymysql==1.1.0
cryptography==42.0.5  # Required for PyMySQL's SHA256 authentication